// src/types/book/BookReview/body.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface BookReviewBody {
  bookShelveId: number
  ratings: number
  review: string
  seq_num: number
  bookInfo: {
    title: string
    author: string
    description: string
    image_url: string
    price: number
    category: string
    publisher: string
    pub_date: string
    isbn: string
    link: string
    [k: string]: unknown
  }
  [k: string]: unknown
}